1) A '.classpath' and '.project' file for eclipse.  Suggestions/additions welcome. 2) A lot of (mostly minor) fixes, tweaks code cleanup thanks to eclipse. CVS: ---------------------------------------------------------------------- CVS: PR: CVS:   If this change addresses a PR in the problem report tracking CVS:   database, then enter the PR number(s) here. CVS: Obtained from: CVS:   If this change has been taken from another system, such as NCSA, CVS:   then name the system in this line, otherwise delete it. CVS: Submitted by: CVS:   If this code has been contributed to Apache by someone else; i.e., CVS:   they sent us a patch or a new module, then include their name/email CVS:   address here. If this is your work then delete this line. CVS: Reviewed by: CVS:   If we are doing pre-commit code reviews and someone else has CVS:   reviewed your changes, include their name(s) here. CVS:   If you have not had it reviewed then delete this line.
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@201272 13f79535-47bb-0310-9956-ffa450edef68

+++ sources/org/apache/batik/bridge/SVGFontUtilities.java
@@ -58,8 +58,6 @@ import java.util.Map;
 import java.util.StringTokenizer;
 
 import org.apache.batik.dom.svg.SVGOMDocument;
-import org.apache.batik.dom.svg.XMLBaseSupport;
-import org.apache.batik.dom.util.XLinkSupport;
 import org.apache.batik.gvt.font.GVTFontFamily;
 import org.apache.batik.gvt.font.GVTFontFace;
 import org.apache.batik.gvt.font.UnresolvedFontFamily;
@@ -70,7 +68,6 @@ import org.apache.batik.css.engine.FontFaceRule;
 
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
-import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
 /**
@@ -223,7 +220,7 @@ public abstract class SVGFontUtilities implements SVGConstants {
                         new StringTokenizer(fontFamilyWeight, " ,");
                     while (st.hasMoreTokens()) {
                         int weightNum = Integer.parseInt(st.nextToken());
-                        int difference = (int)Math.abs(weightNum - i);
+                        int difference = Math.abs(weightNum - i);
                         if (difference < minDifference) {
                             minDifference = difference;
                             minDifferenceIndex = j;
